+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Improved capture-output
 - Improved reporting for `(is (not`, `(is (=` and `(is (not=` for 3+ arguments
+- Show `clojure.data/diff` for `(is (=` and collections
 - `clojure+.core/print-class-tree` #6
 
 # 1.3.1 - Mar 21, 2025